Break logic up into different files and fix imports
